Return just the asset:properties in graphql
