+++ server/graphql/asset/src/types/mod.rs
@@ -1,0 +1,6 @@
+pub mod asset_log;
+pub use asset_log::*;
+pub mod asset;
+pub use asset::*;
+pub mod asset_catalogue_item;
+pub use asset_catalogue_item::*;
